Move LimitedCensusInstance to its own file
